changed directories src/utils,config,models,run*.py->util,config,model,run*.py
