+++ config/config_0005.py
@@ -1,7 +1,8 @@
 import pathlib
 from logging import DEBUG, INFO
+import numpy as np
 
-ROOTDIR = pathlib.Path(__file__).parents[2].resolve()
+ROOTDIR = pathlib.Path(__file__).parents[1].resolve()
 
 
 class RuntimeConfig:
@@ -47,36 +48,35 @@ class ExperimentConfig:
 
 
 class TransformerConfig:
-    features = ['nroman']
+    features = ['magic']
+    # features = ['nroman']
+    DEBUG_SMALL_DATA = False  # use 1% of data if True
 
 
 class ModelConfig:
-    TYPE = 'lgb'
+    TYPE = 'xgb'
+    params = None
 
-    if TYPE == 'lgb':
-        params = {'num_leaves': 491,
-                  'min_child_weight': 0.03454472573214212,
-                  'feature_fraction': 0.3797454081646243,
-                  'bagging_fraction': 0.4181193142567742,
-                  'min_data_in_leaf': 106,
-                  'objective': 'binary',
-                  'max_depth': -1,
-                  'learning_rate': 0.006883242363721497,
-                  "boosting_type": "gbdt",
-                  "bagging_seed": 11,
-                  "metric": 'auc',
-                  "verbosity": -1,
-                  'reg_alpha': 0.3899927210061127,
-                  'reg_lambda': 0.6485237330340494,
-                  'random_state': 47
+    if TYPE == 'xgb':
+        params = {'n_estimators': 2000,
+                  'max_depth': 12,
+                  'learning_rate': 0.02,
+                  'subsample': 0.8,
+                  'colsample_bytree': 0.4,
+                  'missing': np.nan,
+                  'eval_metric': 'auc',
+                  'nthread': 4,
+                  'tree_method': 'hist'
+                  # 'tree_method': 'gpu_hist'
                   }
-    else:
-        raise Exception(f'Model config for {TYPE} is not defined')
+        early_stopping_rounds = 100
+
+    if TYPE is None:
+        raise Exception(f'ModelConfig.params for {TYPE} is not defined')
 
 
 class TrainerConfig:
     model = None
-    early_stopping_rounds = 500
 
     @classmethod
     def set_params(cls, model):
