+++ feature/feature_factory.py
@@ -1,14 +1,8 @@
-from pathlib import Path
-import sys
-
-from features.raw import Raw
-from features.altgor import Altgor
-from features.nroman import Nroman
-from features.magic import Magic
-from utils.mylog import timer
-
-ROOTDIR = Path(__file__).resolve().parents[2]
-sys.path.insert(0, str(ROOTDIR / 'src'))
+from feature.raw import Raw
+from feature.altgor import Altgor
+from feature.nroman import Nroman
+from feature.magic import Magic
+from util.mylog import timer
 
 
 class FeatureFactory:
