+++ feature/nroman.py
@@ -7,12 +7,12 @@ from logging import getLogger
 from sklearn.preprocessing import LabelEncoder
 
 logger = getLogger('main')
-ROOTDIR = Path(__file__).resolve().parents[2]
+ROOTDIR = Path(__file__).resolve().parents[1]
 FEATURE_DIR = ROOTDIR / 'data/feature'
 
-sys.path.insert(0, str(ROOTDIR / 'src'))
-from features.feature_base import Feature
-from utils.mylog import timer
+sys.path.insert(0, str(ROOTDIR))
+from feature.feature_base import Feature
+from util.mylog import timer
 
 
 class Nroman(Feature):
